Separate mimic handling by solver
Use penalty spring forces in SemiImplicit, generalized-coordinate reduction in Featherstone, and keep the XPBD projection local to XPBD and VBD. Perform one VBD mimic projection per solver iteration and move shared kinematic helpers into the simulation module.
