+++ changelog/+joint-mimic-metadata-8e61c4a2.added.md
@@ -1,1 +1,1 @@
-Add joint-owned mimic metadata for joints with matching dimensions, `ModelBuilder.set_joint_mimic()`, componentwise follower-coordinate updates through `eval_mimic()`, solver support in Featherstone, SemiImplicit, XPBD, MuJoCo, and VBD, and a lead-screw example.
+Add joint-owned mimic metadata for non-quaternion joints with matching dimensions, `ModelBuilder.set_joint_mimic()`, componentwise follower-coordinate updates through `eval_mimic()`, solver support in Featherstone, SemiImplicit, XPBD, MuJoCo, and VBD, and a lead-screw example. SemiImplicit enforces mimic relationships as penalty springs with global `joint_mimic_ke` and `joint_mimic_kd` gains.
