+++ docs/solvers/index.rst
@@ -326,7 +326,7 @@ constraints, with opt-in unified compliant ALM and a deprecated legacy AVBD path
      - |yes| :sup:`2`
      - |no|
      - |yes|
-     - |yes| :sup:`6`
+     - |yes| :sup:`7`
      - |no|
    * - :attr:`~newton.Model.joint_effort_limit`
      - |no|
@@ -364,7 +364,7 @@ constraints, with opt-in unified compliant ALM and a deprecated legacy AVBD path
      - |yes| :sup:`2`
      - |yes|
      - |yes|
-     - |yes| :sup:`6`
+     - |yes| :sup:`7`
      - |yes|
    * - :attr:`~newton.Model.joint_target_mode`
      - |no|
@@ -402,18 +402,19 @@ constraints, with opt-in unified compliant ALM and a deprecated legacy AVBD path
      - |yes|
      - |no|
      - |no|
-   * - Mimic constraints
+   * - Mimic joints
      - |yes| :sup:`3`
      - |yes| :sup:`4`
-     - |yes| :sup:`4`
      - |yes| :sup:`5`
-     - |yes| :sup:`4`
+     - |yes| :sup:`6`
+     - |yes| :sup:`5`
      - |no|
 
 | :sup:`3` Featherstone eliminates follower degrees of freedom from its reduced dynamics and transfers follower forces and inertia to the reference joint.
-| :sup:`4` SemiImplicit, XPBD, and VBD enforce joint-owned mimic relationships through coupled maximal-coordinate corrections.
-| :sup:`5` MuJoCo lowers each joint-owned relationship to joint equality constraints. Multi-axis D6 relationships produce one equality constraint per axis.
-| :sup:`6` VBD interprets ``joint_target_kd`` and ``joint_limit_kd`` as absolute damping coefficients in physical units.
+| :sup:`4` SemiImplicit enforces joint-owned mimic relationships with penalty springs configured by ``joint_mimic_ke`` and ``joint_mimic_kd``.
+| :sup:`5` XPBD and VBD enforce joint-owned mimic relationships through coupled maximal-coordinate corrections. Both apply one mimic correction per solver iteration.
+| :sup:`6` MuJoCo lowers each joint-owned relationship to joint equality constraints. Multi-axis D6 relationships produce one equality constraint per axis.
+| :sup:`7` VBD interprets ``joint_target_kd`` and ``joint_limit_kd`` as absolute damping coefficients in physical units.
