+++ newton/_src/sim/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025 The Newton Developers
 # SPDX-License-Identifier: Apache-2.0
 
-from .articulation import eval_fk, eval_ik, eval_inverse_dynamics_force, eval_jacobian, eval_mass_matrix, eval_mimic
+from .articulation import eval_fk, eval_ik, eval_inverse_dynamics_force, eval_jacobian, eval_mass_matrix
 from .builder import ModelBuilder
 from .collide import CollisionPipeline
 from .contact_kinematics import eval_rigid_contact_kinematics
@@ -16,6 +16,7 @@ from .enums import (
     StateFlags,
 )
 from .inverse_dynamics import eval_inverse_dynamics_passive
+from .joint_mimic import eval_mimic
 from .model import Model
 from .rod import Rod
 from .state import State
