+++ newton/_src/sim/articulation.py
@@ -669,83 +669,6 @@ def eval_fk(
 
 
 @wp.kernel
-def eval_mimic_joints(
-    joint_mimic_joint: wp.array[int],
-    joint_mimic_coeffs: wp.array[wp.vec2],
-    joint_q_start: wp.array[int],
-    joint_qd_start: wp.array[int],
-    # outputs
-    joint_q: wp.array[float],
-    joint_qd: wp.array[float],
-):
-    joint = wp.tid()
-    reference_joint = joint_mimic_joint[joint]
-    if reference_joint < 0:
-        return
-
-    coeffs = joint_mimic_coeffs[joint]
-    q_start = joint_q_start[joint]
-    reference_q_start = joint_q_start[reference_joint]
-    for coordinate in range(joint_q_start[joint + 1] - q_start):
-        joint_q[q_start + coordinate] = coeffs[0] + coeffs[1] * joint_q[reference_q_start + coordinate]
-
-    qd_start = joint_qd_start[joint]
-    reference_qd_start = joint_qd_start[reference_joint]
-    for dof in range(joint_qd_start[joint + 1] - qd_start):
-        joint_qd[qd_start + dof] = coeffs[1] * joint_qd[reference_qd_start + dof]
-
-
-def eval_mimic(model: Model, state_in: State, state_out: State | None = None) -> None:
-    """Update follower joint coordinates from their reference joints.
-
-    For each follower, this function reads every position and velocity
-    coordinate of the reference joint, then writes the matching follower
-    coordinates according to :attr:`Model.joint_mimic_coeffs`. Independent
-    joints are left unchanged. Only :attr:`State.joint_q` and
-    :attr:`State.joint_qd` are written.
-
-    If ``state_out`` is omitted, ``state_in`` is updated in place. Otherwise,
-    all joint coordinates are first copied from ``state_in`` to ``state_out``
-    and the followers are updated in ``state_out``.
-
-    Args:
-        model: Model containing the joint mimic metadata.
-        state_in: State providing the input joint coordinates.
-        state_out: State receiving the updated joint coordinates. If ``None``,
-            update ``state_in`` in place.
-
-    Raises:
-        ValueError: If either state does not contain joint coordinate arrays.
-    """
-    if state_in.joint_q is None or state_in.joint_qd is None:
-        raise ValueError("state_in must contain joint_q and joint_qd arrays")
-
-    if state_out is None:
-        state_out = state_in
-    elif state_out.joint_q is None or state_out.joint_qd is None:
-        raise ValueError("state_out must contain joint_q and joint_qd arrays")
-    elif state_out is not state_in:
-        state_out.joint_q.assign(state_in.joint_q)
-        state_out.joint_qd.assign(state_in.joint_qd)
-
-    if model.joint_count == 0:
-        return
-
-    wp.launch(
-        kernel=eval_mimic_joints,
-        dim=model.joint_count,
-        inputs=[
-            model.joint_mimic_joint,
-            model.joint_mimic_coeffs,
-            model.joint_q_start,
-            model.joint_qd_start,
-        ],
-        outputs=[state_out.joint_q, state_out.joint_qd],
-        device=model.device,
-    )
-
-
-@wp.kernel
 def compute_shape_world_transforms(
     shape_transform: wp.array[wp.transform],
     shape_body: wp.array[int],
