+++ newton/_src/solvers/semi_implicit/kernels_body.py
@@ -12,6 +12,7 @@ from ...sim import (
     Model,
     State,
 )
+from ...sim.joint_mimic import eval_joint_mimic_coordinate, eval_joint_mimic_velocity
 
 
 @wp.func
@@ -51,6 +52,115 @@ def joint_force(
     return limit_f + damping_f + target_f + passive_f
 
 
+@wp.func
+def eval_joint_mimic_forces(
+    follower: int,
+    body_q: wp.array[wp.transform],
+    body_qd: wp.array[wp.spatial_vector],
+    body_com: wp.array[wp.vec3],
+    joint_qd_start: wp.array[int],
+    joint_type: wp.array[int],
+    joint_enabled: wp.array[bool],
+    joint_child: wp.array[int],
+    joint_parent: wp.array[int],
+    joint_X_p: wp.array[wp.transform],
+    joint_X_c: wp.array[wp.transform],
+    joint_axis: wp.array[wp.vec3],
+    joint_dof_dim: wp.array2d[int],
+    joint_mimic_joint: wp.array[int],
+    joint_mimic_coeffs: wp.array[wp.vec2],
+    joint_mimic_ke: float,
+    joint_mimic_kd: float,
+    body_f: wp.array[wp.spatial_vector],
+):
+    """Accumulate spring-damper forces for one joint-owned mimic relationship."""
+    reference = joint_mimic_joint[follower]
+    if reference < 0 or not joint_enabled[reference]:
+        return
+
+    follower_type = joint_type[follower]
+    reference_type = joint_type[reference]
+    follower_supported = (
+        follower_type == JointType.PRISMATIC or follower_type == JointType.REVOLUTE or follower_type == JointType.D6
+    )
+    reference_supported = (
+        reference_type == JointType.PRISMATIC or reference_type == JointType.REVOLUTE or reference_type == JointType.D6
+    )
+    if not follower_supported or not reference_supported:
+        return
+
+    follower_parent = joint_parent[follower]
+    follower_child = joint_child[follower]
+    reference_parent = joint_parent[reference]
+    reference_child = joint_child[reference]
+    coordinate_count = joint_dof_dim[follower, 0] + joint_dof_dim[follower, 1]
+    follower_linear_count = joint_dof_dim[follower, 0]
+    reference_linear_count = joint_dof_dim[reference, 0]
+    coeffs = joint_mimic_coeffs[follower]
+    offset = coeffs[0]
+    multiplier = coeffs[1]
+
+    for component in range(coordinate_count):
+        follower_q, follower_parent_gradient, follower_child_gradient = eval_joint_mimic_coordinate(
+            follower,
+            component,
+            body_q,
+            body_com,
+            joint_type,
+            joint_parent,
+            joint_child,
+            joint_X_p,
+            joint_X_c,
+            joint_qd_start,
+            joint_dof_dim,
+            joint_axis,
+        )
+        reference_q, reference_parent_gradient, reference_child_gradient = eval_joint_mimic_coordinate(
+            reference,
+            component,
+            body_q,
+            body_com,
+            joint_type,
+            joint_parent,
+            joint_child,
+            joint_X_p,
+            joint_X_c,
+            joint_qd_start,
+            joint_dof_dim,
+            joint_axis,
+        )
+
+        position_error = follower_q - offset - multiplier * reference_q
+        if component >= follower_linear_count and component >= reference_linear_count:
+            position_error = wp.atan2(wp.sin(position_error), wp.cos(position_error))
+
+        follower_qd = eval_joint_mimic_velocity(
+            follower_parent,
+            follower_child,
+            follower_parent_gradient,
+            follower_child_gradient,
+            body_qd,
+        )
+        reference_qd = eval_joint_mimic_velocity(
+            reference_parent,
+            reference_child,
+            reference_parent_gradient,
+            reference_child_gradient,
+            body_qd,
+        )
+        velocity_error = follower_qd - multiplier * reference_qd
+        force = -joint_mimic_ke * position_error - joint_mimic_kd * velocity_error
+
+        if follower_parent >= 0:
+            wp.atomic_add(body_f, follower_parent, follower_parent_gradient * force)
+        if follower_child >= 0:
+            wp.atomic_add(body_f, follower_child, follower_child_gradient * force)
+        if reference_parent >= 0:
+            wp.atomic_add(body_f, reference_parent, reference_parent_gradient * (-multiplier * force))
+        if reference_child >= 0:
+            wp.atomic_add(body_f, reference_child, reference_child_gradient * (-multiplier * force))
+
+
 @wp.kernel
 def eval_body_joints(
     body_q: wp.array[wp.transform],
@@ -76,8 +186,12 @@ def eval_body_joints(
     joint_limit_ke: wp.array[float],
     joint_limit_kd: wp.array[float],
     joint_damping: wp.array[float],
+    joint_mimic_joint: wp.array[int],
+    joint_mimic_coeffs: wp.array[wp.vec2],
     joint_attach_ke: float,
     joint_attach_kd: float,
+    joint_mimic_ke: float,
+    joint_mimic_kd: float,
     body_f: wp.array[wp.spatial_vector],
 ):
     tid = wp.tid()
@@ -513,7 +627,28 @@ def eval_body_joints(
                 )
             )
 
-    # write forces
+    eval_joint_mimic_forces(
+        tid,
+        body_q,
+        body_qd,
+        body_com,
+        joint_qd_start,
+        joint_type,
+        joint_enabled,
+        joint_child,
+        joint_parent,
+        joint_X_p,
+        joint_X_c,
+        joint_axis,
+        joint_dof_dim,
+        joint_mimic_joint,
+        joint_mimic_coeffs,
+        joint_mimic_ke,
+        joint_mimic_kd,
+        body_f,
+    )
+
+    # write joint forces
     if c_parent >= 0:
         wp.atomic_add(body_f, c_parent, wp.spatial_vector(f_total, t_total + wp.cross(r_p, f_total)))
 
@@ -521,7 +656,14 @@ def eval_body_joints(
 
 
 def eval_body_joint_forces(
-    model: Model, state: State, control: Control, body_f: wp.array, joint_attach_ke: float, joint_attach_kd: float
+    model: Model,
+    state: State,
+    control: Control,
+    body_f: wp.array,
+    joint_attach_ke: float,
+    joint_attach_kd: float,
+    joint_mimic_ke: float,
+    joint_mimic_kd: float,
 ):
     if model.joint_count:
         wp.launch(
@@ -551,8 +693,12 @@ def eval_body_joint_forces(
                 model.joint_limit_ke,
                 model.joint_limit_kd,
                 model.joint_damping,
+                model.joint_mimic_joint,
+                model.joint_mimic_coeffs,
                 joint_attach_ke,
                 joint_attach_kd,
+                joint_mimic_ke,
+                joint_mimic_kd,
             ],
             outputs=[body_f],
             device=model.device,
